Finished

+++ gui/car-dashboard/hal/esp32/app_hal.h
@@ -9,8 +9,8 @@ extern "C" {
 void hal_setup(void);
 void hal_loop(void);
 
-int hal_get_altitude(void);
-int hal_get_pitch(void);
+float hal_get_altitude(void);
+float hal_get_pitch(void);
 
 #ifdef __cplusplus
 } /* extern "C" */
