+++ gui/car-dashboard/hal/sdl2/app_hal.h
@@ -7,8 +7,8 @@ extern "C" {
 void hal_setup(void);
 void hal_loop(void);
 
-int hal_get_altitude(void);
-int hal_get_pitch(void);
+float hal_get_altitude(void);
+float hal_get_pitch(void);
 
 void delay(unsigned long);
